requirements.txt
python-dotenv
openai
pandas
numpy
tqdm